+++ ydb/core/base/events.h
@@ -174,6 +174,12 @@ struct TKikimrEvents : TEvents {
         ES_REPLICATION_SERVICE,
         ES_CHANGE_EXCHANGE,
         ES_S3_FILE_QUEUE,
+        ES_S3_PROVIDER,
+        ES_NEBIUS_ACCESS_SERVICE,
+        ES_BACKUP_SERVICE,
+        ES_TX_BACKGROUND,
+        ES_SS_BG_TASKS,
+        ES_LIMITER
     };
 };
