+++ ydb/library/yql/providers/s3/provider/yql_s3_settings.cpp
@@ -26,6 +26,7 @@ TS3Configuration::TS3Configuration()
     REGISTER_SETTING(*this, FileQueueBatchSizeLimit);
     REGISTER_SETTING(*this, FileQueueBatchObjectCountLimit);
     REGISTER_SETTING(*this, FileQueuePrefetchSize);
+    REGISTER_SETTING(*this, AsyncDecoding);
 }
 
 TS3Settings::TConstPtr TS3Configuration::Snapshot() const {
