+++ ydb/library/yql/providers/s3/ya.make
@@ -3,6 +3,7 @@ RECURSE(
     common
     compressors
     credentials
+    events
     expr_nodes
     object_listers
     path_generator
